Refactor test suite
Extract all test cases
