+++ cli/lib/simplification/compact-triangles.test.js
@@ -1,0 +1,12 @@
+import compactTriangles from "./compact-triangles.js";
+
+describe("compactTriangles", () => {
+  it("should remove deleted triangles", () => {
+    let newTriangles = compactTriangles([
+      { deleted: true },
+      { deleted: true },
+      { deleted: false },
+    ]);
+    expect(newTriangles.length).toBe(1);
+  });
+});
